dreamhack/web/dreamboard/requirements.txt
falcon==2.0.0
falcon-auth==1.1.0
falcon-cors==1.1.7
gunicorn==20.0.4
PyJWT==1.7.1
requests==2.24.0
selenium==3.141.0
SQLAlchemy==1.3.19
